refactor: improve vector stores code structure and maintainability
- Extract shared types to dedicated types.ts file
- Separate status cards into standalone components
- Create dedicated repository list component
- Extract data fetching utilities to data.ts
- Simplify main page component by removing code duplication
- Improve error handling with try-catch
- Remove redundant type definitions across files
- Maintain exact same UI/UX while improving code organization
- Reduce bundle size through better component separation
- Enhance maintainability and testability
